2/requirements.txt
numpy
pytorch
torchvision
